requirements.txt
streamlit
openai
langchain
langchain_community
langchain_openai
langchain_core
msal
psycopg2-binary
duckduckgo-search
langchain_postgres
pypdf
streamlit_extras